+++ reqcoco-parser-aggregator/src/main/java/net/paissad/tools/reqcoco/parser/aggregator/AggregatorReqDeclParser.java
@@ -9,6 +9,7 @@ import java.nio.file.Paths;
 import java.nio.file.SimpleFileVisitor;
 import java.nio.file.attribute.BasicFileAttributes;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
@@ -78,7 +79,7 @@ public class AggregatorReqDeclParser extends AbstractReqDeclParser {
 
     @Override
     public Collection<String> getRegisteredFileExtensions() {
-        return null;
+        return Collections.emptyList();
     }
 
     private static String getFileExtension(final String filename) {