+++ reqcoco-parser-redmine/src/test/java/net/paissad/tools/reqcoco/parser/redmine/RedmineReqSourceParserTest.java
@@ -52,7 +52,7 @@ public class RedmineReqSourceParserTest {
 
     @Test
     public void testgetRegisteredFileExtensions() {
-        Assert.assertNull(this.redmineReqSourceParser.getRegisteredFileExtensions());
+        Assert.assertTrue(this.redmineReqSourceParser.getRegisteredFileExtensions().isEmpty());
     }
 
 	@Test