+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Unreleased]
 ### Added
  - Retrieve requirements declarations recursively from files of a directory (issue #20)
+ - Add Aggregator Parser (issue #21)
  - Generate JSON output coverage report (issue #22)
 
 ## [1.0.3] - 2018-09-19